Changed to STM32CubeFx. Much better. But still problems!
The .h for the F103 (the ST-LINK) is very strange. No comments, the GPIO
typedefs have no entries for the MODER etc. The comments with the
offsets for OB are wrong, or seem to be. They don't match the offsets
for the other chips...

The F4 .h (for F4 disco board) lacks USB info, even though the chip
totally supports it.

CAN is still broken. HRTIM is broken.

Parsing with patterns is a true PITA. But ATM, except for the embedded
typedef problem (CAN and HRTIM), I'm getting pretty good results.