Parse out exceptions and print them
Unsure if I want to keep the exception numbers as they are (ie, -16 to
-1 are Cortex-M exceptions, and 0 on up are chip-specific exceptions) or
offset them by 16, so 0 to 15 are Cortex-M, and the chip-specific ones
start at 16.

I'll decide which seems to work better when I start writing code to
write handler addresses into a vector table.

Also: do the names need to be changed? On S08 they are V<name> to show
that they are vectors.