Now generates register bitfields too! (and finds errors in .h files)
I am inadvertently "linting" ST's .h files by doing this.

Instead of simply copying .h files out of an STM32Cube zip file
converting it, I'm either going to have to check in the .h files I care
about (with changes to fix the brokenness and typos) or keep a set of
patch files.

Right now the registers are all printing, then the bitfields are all
printed. They are *not* interleaved. (I may tinker with ways to make
this happen.)

Also, the second part of each field definition is a *mask* rather than a
*width*. This is so in part because *one* of the masks defined in
stm32l152xb.h (RI_ASCR1_CH_Msk) is defined as 0x3FCFFFF - which has gaps
in it and can't be represented simply by a width! It should really be an
*equate* and not a mask.

I've changed the way registers are printed as well. They used to be
<peripheral>.<register> but since all the fields are defined as
<periph>_<register> I switched to that convention.

I could have changed the fields instead. I'm not sure which I prefer!

periph.reg is more like C.