Big change: removed all downloaded .h files; getting from github
ST has decided to publish, in a modular way, the pieces that make the
STM32Cube downloads.

While it used to be necessary to download multi-megabyte ZIP files just
to extract the interesting .h files, now it's possible to grab much
smaller archives that represent the master branch of the "cmsis" files -
basically the .h files and some .s C startup files.

This makes our job much easier!

Now you just do

  make download all

to download the latest master branches and generate first a handful of
.h files, and from these a handful of .mu4 (muforth) files.